Add index.html and game.js
Add a blank html document linked to a javascript file. The game.js have two blank functions and one computerPlay function that returns a random move; It is accomplished by
creating an array with the moves as its items and accesing the index of the array with an random number generator.