Add titleCase function and playRound function
titleCase is for making the user input case-insensitive. The playRound function accepts two parameters and is used to compare the result of the round. Furthermore, the playRound function checks whether the user input is not a truthy value; If it is, it returns a string. Otherwise, it proceeds to the code below.